requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
openai==1.72.0
python-dotenv==1.1.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.1
pydantic_core==2.33.0
certifi==2025.1.31
idna==3.10
sniffio==1.3.1
anyio==4.9.0
starlette==0.46.1
